ui/requirements.txt
Werkzeug==2.0.3
dash_auth
dash_bootstrap_components
folium==0.14.0
geopy==2.3.0
loguru==0.6.0
numpy==1.23.5
openpyxl==3.0.10
openrouteservice==2.3.3
pandas==1.5.2
requests==2.28.1
s3fs==2022.11.0
python-dotenv
loguru